Issue 67.1: extract session-delta cluster from afxdp.rs
afxdp.rs goes from 1,454 → 1,275 prod LOC by relocating three
session-delta-processing fns into a new sibling module
afxdp/session_delta.rs:

- flush_session_deltas (165 LOC) — drains per-binding SessionDelta
  queues, applies them to the SessionTable, emits the corresponding
  events to the event-stream channel
- purge_queued_flows_for_closed_deltas (12 LOC) — post-processes the
  pending-forward queues to drop frames for closed flows
- session_delta_event (7 LOC) — SessionDeltaKind → &'static str

Visibility widened from file-private `fn` to `pub(super) fn` so
afxdp.rs's `use session_delta::{...};` resolves.

Pure relocation: bodies byte-for-byte identical, all types and
helpers reachable through `use super::*;` from the child module.

cargo test --release: 865 passed, 0 failed.
