+++ userspace-dp/src/afxdp.rs
@@ -678,6 +678,11 @@ fn poll_binding(
 
 mod poll_descriptor;
 use poll_descriptor::poll_binding_process_descriptor;
+
+// Issue 67.1: session-delta processing (flush_session_deltas et al.)
+// extracted into afxdp/session_delta.rs.
+mod session_delta;
+use session_delta::{flush_session_deltas, purge_queued_flows_for_closed_deltas};
 fn retry_pending_neigh(
     binding: &mut BindingWorker,
     left: &mut [BindingWorker],
@@ -1035,190 +1040,6 @@ fn find_target_binding_mut<'a>(
     binding_by_index_mut(left, current_index, ingress_binding, right, target_index)
 }
 
-fn purge_queued_flows_for_closed_deltas(bindings: &mut [BindingWorker], deltas: &[SessionDelta]) {
-    for delta in deltas {
-        if delta.kind != SessionDeltaKind::Close {
-            continue;
-        }
-        let reverse_key = reverse_session_key(&delta.key, delta.decision.nat);
-        for binding in bindings.iter_mut() {
-            cancel_queued_flow_on_binding(binding, &delta.key, &reverse_key);
-        }
-    }
-}
-
-fn flush_session_deltas(
-    ident: &BindingIdentity,
-    live: &BindingLiveState,
-    session_map_fd: c_int,
-    conntrack_v4_fd: c_int,
-    conntrack_v6_fd: c_int,
-    deltas: &[SessionDelta],
-    shared_sessions: &Arc<Mutex<FastMap<SessionKey, SyncedSessionEntry>>>,
-    shared_nat_sessions: &Arc<Mutex<FastMap<SessionKey, SyncedSessionEntry>>>,
-    shared_forward_wire_sessions: &Arc<Mutex<FastMap<SessionKey, SyncedSessionEntry>>>,
-    shared_owner_rg_indexes: &SharedSessionOwnerRgIndexes,
-    recent_session_deltas: &Arc<Mutex<VecDeque<SessionDeltaInfo>>>,
-    peer_worker_commands: &[Arc<Mutex<VecDeque<WorkerCommand>>>],
-    event_stream: &Option<crate::event_stream::EventStreamWorkerHandle>,
-    forwarding: &ForwardingState,
-) {
-    let zone_name_to_id = &forwarding.zone_name_to_id;
-    let zone_id_to_name = &forwarding.zone_id_to_name;
-    for delta in deltas {
-        // #919/#922: emit both the resolved zone NAMES (legacy field,
-        // empty when the ID is unknown) and the u16 IDs. New daemons
-        // prefer the IDs; older daemons read the names. The previous
-        // code wrote `metadata.ingress_zone.to_string()` here, which
-        // produced "1"/"2" string literals that broke `zoneIDs[name]`
-        // on the Go side.
-        let ingress_name = zone_id_to_name
-            .get(&delta.metadata.ingress_zone)
-            .cloned()
-            .unwrap_or_default();
-        let egress_name = zone_id_to_name
-            .get(&delta.metadata.egress_zone)
-            .cloned()
-            .unwrap_or_default();
-        let info = SessionDeltaInfo {
-            timestamp: Utc::now(),
-            slot: ident.slot,
-            queue_id: ident.queue_id,
-            worker_id: ident.worker_id,
-            interface: ident.interface.to_string(),
-            ifindex: ident.ifindex,
-            event: session_delta_event(delta.kind).to_string(),
-            addr_family: delta.key.addr_family,
-            protocol: delta.key.protocol,
-            src_ip: delta.key.src_ip.to_string(),
-            dst_ip: delta.key.dst_ip.to_string(),
-            src_port: delta.key.src_port,
-            dst_port: delta.key.dst_port,
-            ingress_zone: ingress_name,
-            egress_zone: egress_name,
-            ingress_zone_id: delta.metadata.ingress_zone,
-            egress_zone_id: delta.metadata.egress_zone,
-            owner_rg_id: delta.metadata.owner_rg_id,
-            disposition: match delta.decision.resolution.disposition {
-                ForwardingDisposition::ForwardCandidate => "forward_candidate",
-                ForwardingDisposition::LocalDelivery => "local_delivery",
-                ForwardingDisposition::NoRoute => "no_route",
-                ForwardingDisposition::MissingNeighbor => "missing_neighbor",
-                ForwardingDisposition::PolicyDenied => "policy_denied",
-                ForwardingDisposition::FabricRedirect => "fabric_redirect",
-                ForwardingDisposition::HAInactive => "ha_inactive",
-                ForwardingDisposition::DiscardRoute => "discard_route",
-                ForwardingDisposition::NextTableUnsupported => "next_table_unsupported",
-            }
-            .to_string(),
-            origin: delta.origin.as_str().to_string(),
-            egress_ifindex: delta.decision.resolution.egress_ifindex,
-            tx_ifindex: delta.decision.resolution.tx_ifindex,
-            tunnel_endpoint_id: delta.decision.resolution.tunnel_endpoint_id,
-            tx_vlan_id: delta.decision.resolution.tx_vlan_id,
-            next_hop: delta
-                .decision
-                .resolution
-                .next_hop
-                .map(|ip| ip.to_string())
-                .unwrap_or_default(),
-            neighbor_mac: delta
-                .decision
-                .resolution
-                .neighbor_mac
-                .map(format_mac)
-                .unwrap_or_default(),
-            src_mac: delta
-                .decision
-                .resolution
-                .src_mac
-                .map(format_mac)
-                .unwrap_or_default(),
-            nat_src_ip: delta
-                .decision
-                .nat
-                .rewrite_src
-                .map(|ip| ip.to_string())
-                .unwrap_or_default(),
-            nat_dst_ip: delta
-                .decision
-                .nat
-                .rewrite_dst
-                .map(|ip| ip.to_string())
-                .unwrap_or_default(),
-            nat_src_port: delta.decision.nat.rewrite_src_port.unwrap_or(0),
-            nat_dst_port: delta.decision.nat.rewrite_dst_port.unwrap_or(0),
-            fabric_redirect: delta.fabric_redirect_sync
-                || delta.decision.resolution.disposition == ForwardingDisposition::FabricRedirect,
-            fabric_ingress: delta.metadata.fabric_ingress,
-        };
-        live.push_session_delta(info.clone());
-        // Push to event stream (new path) alongside existing RPC fallback.
-        if let Some(es) = event_stream {
-            es.push_delta(delta, zone_name_to_id);
-        }
-        if let Ok(mut recent) = recent_session_deltas.lock() {
-            push_recent_session_delta(&mut recent, info);
-        }
-        if delta.kind == SessionDeltaKind::Close {
-            if cfg!(feature = "debug-log") {
-                debug_log!(
-                    "SESS_DELETE: proto={} {}:{} -> {}:{} nat_src={:?} nat_dst={:?} bpf_entries_before={}",
-                    delta.key.protocol,
-                    delta.key.src_ip,
-                    delta.key.src_port,
-                    delta.key.dst_ip,
-                    delta.key.dst_port,
-                    delta.decision.nat.rewrite_src,
-                    delta.decision.nat.rewrite_dst,
-                    count_bpf_session_entries(session_map_fd),
-                );
-            }
-            delete_live_session_entry(
-                session_map_fd,
-                &delta.key,
-                delta.decision.nat,
-                delta.metadata.is_reverse,
-            );
-            delete_bpf_conntrack_entry(conntrack_v4_fd, conntrack_v6_fd, &delta.key);
-            remove_shared_session(
-                shared_sessions,
-                shared_nat_sessions,
-                shared_forward_wire_sessions,
-                &shared_owner_rg_indexes,
-                &delta.key,
-            );
-            let reverse_key = reverse_session_key(&delta.key, delta.decision.nat);
-            delete_live_session_entry(session_map_fd, &reverse_key, delta.decision.nat, true);
-            delete_bpf_conntrack_entry(conntrack_v4_fd, conntrack_v6_fd, &reverse_key);
-            remove_shared_session(
-                shared_sessions,
-                shared_nat_sessions,
-                shared_forward_wire_sessions,
-                &shared_owner_rg_indexes,
-                &reverse_key,
-            );
-            replicate_session_delete(peer_worker_commands, &delta.key);
-            replicate_session_delete(
-                peer_worker_commands,
-                &reverse_session_key(&delta.key, delta.decision.nat),
-            );
-            if cfg!(feature = "debug-log") {
-                debug_log!(
-                    "SESS_DELETE_DONE: bpf_entries_after={}",
-                    count_bpf_session_entries(session_map_fd),
-                );
-            }
-        }
-    }
-}
-
-fn session_delta_event(kind: SessionDeltaKind) -> &'static str {
-    match kind {
-        SessionDeltaKind::Open => "open",
-        SessionDeltaKind::Close => "close",
-    }
-}
 
 fn record_exception(
     recent_exceptions: &Arc<Mutex<VecDeque<ExceptionStatus>>>,
